pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "matplotlib>=3.10.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
